+++ activemodel/lib/active_model/validations/callbacks.rb
@@ -106,6 +106,16 @@ module ActiveModel
                 *options[:if]
               ]
             end
+
+            if options.key?(:except_on)
+              options[:except_on] = Array(options[:except_on])
+              options[:unless] = [
+                ->(o) {
+                  options[:except_on].intersect?(Array(o.validation_context))
+                },
+                *options[:unless]
+              ]
+            end
           end
       end